fixed something in the check syntax

+++ tokenization_utils1.c
@@ -6,7 +6,7 @@
 /*   By: jpelaez- <jpelaez-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/30 12:22:27 by jpelaez-          #+#    #+#             */
-/*   Updated: 2023/06/30 12:35:58 by jpelaez-         ###   ########.fr       */
+/*   Updated: 2023/07/05 15:58:44 by jpelaez-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -36,18 +36,18 @@ int	count_tokens(char *line, char *del)
 	return (len);
 }
 
-int check_delimiter(char c, char *delimiter)
+int	check_delimiter(char c, char *delimiter)
 {
-    size_t	j;
+	int j;
 
 	j = 0;
 	if (ft_strlen(delimiter) == 0)
-		return (0);	
-    while (delimiter[j] != c && delimiter[j] != '\0')
-    {
-        if(delimiter[j] == c)
-            return(1);
-        j++;
-    }
+		return (0);
+	while (delimiter[j] != c && delimiter[j] != '\0')
+	{
+		if (delimiter[j] == c)
+			return (1);
+		j++;
+	}
 	return (0);
 }
